requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.10
langchain-openai==0.1.0
python-dotenv==1.0.0
supabase==2.7.4
pymupdf==1.24.0
python-docx==1.1.0
httpx>=0.27.0
